Style

+++ Source/Luft/Base.hpp
@@ -10,7 +10,7 @@ using uint16 = unsigned short;
 using uint32 = unsigned int;
 using uint64 = unsigned long long;
 
-using usize = unsigned long long;
+using usize = uint64;
 
 using bool32 = uint32;
